feat: 阶段一开发完成 - 重构首页为Landing Page和创建Dashboard转换工具
